safari-534.53-branch/LayoutTests/svg/dynamic-updates/SVGFEDisplacementMapElement-svgdom-in-prop-expected.txt

CONSOLE MESSAGE: line 51: TypeError: 'undefined' is not an object (evaluating 'displacementMap.in1.baseVal = "Texture"')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' property of the SVGFEDisplacementMapElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL displacementMap.in1.baseVal should be Map. Threw exception TypeError: 'undefined' is not an object (evaluating 'displacementMap.in1.baseVal')
